+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- Added a flag (step_on_batch) to the LR Scheduler callbacks which allows for step() to be called on each iteration instead of each epoch
 ### Changed
 - Changed the order of the arguments to the lambda function in the EpochLambda metric for consistency with pytorch and other metrics
 ### Deprecated